requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 大模型相关
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10

# 向量数据库和检索
pymilvus==2.3.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# 知识图谱
neo4j==5.14.0
py2neo==2021.2.4

# 数据处理
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pdfplumber==0.10.3
python-docx==1.1.0
pypdf2==3.0.1

# OCR和图像处理
pytesseract==0.3.10
pillow==10.1.0
opencv-python==4.8.1.78

# 缓存和数据库
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# 工具库
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
cryptography==41.0.7  # AES加密

# 异步HTTP客户端
httpx==0.25.1
aiohttp==3.9.1

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# 语音处理
SpeechRecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1

# 模型训练和微调
torch==2.1.0
transformers==4.35.2
peft==0.7.1  # LoRA微调
datasets==2.14.7
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
shap==0.43.0  # 特征重要性分析

# 其他工具
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1

